pyproject.toml
[tool.poetry]
name = "omni-assignment"
version = "0.1.0"
description = ""
authors = ["William Cher"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
flake8-isort = "^6.1.1"
coverage = "^7.6.1"
black = "^24.8.0"
pre-commit = "^3.8.0"
pytest = "^8.3.2"
factory-boy = "^3.3.1"
django-coverage-plugin = "^3.1.0"
pytest-django = "^4.8.0"


[tool.poetry.group.production.dependencies]
psycopg = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
